python-files/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gpiod==2.3.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lirc==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
ncnn==1.0.20240820
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
RPi.GPIO==0.7.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.18
ultralytics-thop==2.0.9
urllib3==2.2.3
VimbaPython==1.3.0
